requirements.txt
docker_py==1.10.6
Requests==2.31.0
schedule==1.2.1
